Factor out the commonly used components in both aggregate
