+++ src/operators/aggregate.h
@@ -28,27 +28,10 @@
 #include "parallel_hashmap/phmap.h"
 #include "storage/block.h"
 #include "storage/table.h"
+#include "aggregate_const.h"
 
 namespace hustle {
 namespace operators {
-
-// Types of aggregates we can perform. COUNT is currently not supported.
-enum AggregateKernel { SUM, COUNT, MEAN };
-
-/**
- * A reference structure containing all the information needed to perform an
- * aggregate over a column.
- *
- * @param kernel The type of aggregate we want to compute
- * @param agg_name Name of the new aggregate column
- * @param col_ref A reference to the column over which we want to compute the
- * aggregate
- */
-struct AggregateReference {
-  AggregateKernel kernel;
-  std::string agg_name;
-  ColumnReference col_ref;
-};
 
 /**
  * Group = a set of column values, one for each column in the GROUP BY clause
